requirements.txt
opencv-contrib-python
numpy
matplotlib
tensorflow
keras
python-socketio
Flask
eventlet
python-engineio==3.13.2
python-socketio==4.6.1